# smartmeeting-api/swagger.json
{"openapi":"3.0.1","info":{"title":"SmartMeeting API","description":"API para gerenciamento de reuniões, salas e participantes","contact":{"name":"Equipe SmartMeeting","url":"https://smartmeeting.com","email":"contato@smartmeeting.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"JWT":[]}],"paths":{"/tarefas/{id}":{"get":{"tags":["tarefa-controller"],"operationId":"buscarPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TarefaDTO"}}}}}},"put":{"tags":["tarefa-controller"],"operationId":"atualizar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TarefaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TarefaDTO"}}}}}},"delete":{"tags":["tarefa-controller"],"operationId":"deletar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/salas/{id}":{"get":{"tags":["sala-controller"],"operationId":"buscarPorId_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SalaDTO"}}}}}},"put":{"tags":["sala-controller"],"operationId":"atualizar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SalaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SalaDTO"}}}}}},"delete":{"tags":["sala-controller"],"operationId":"deletar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/reunioes/{id}":{"get":{"tags":["reuniao-controller"],"operationId":"buscarPorId_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReuniaoDTO"}}}}}},"put":{"tags":["reuniao-controller"],"operationId":"atualizar_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReuniaoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReuniaoDTO"}}}}}},"delete":{"tags":["reuniao-controller"],"operationId":"deletar_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/pessoas/{id}":{"get":{"tags":["pessoa-controller"],"operationId":"buscarPorId_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PessoaDTO"}}}}}},"put":{"tags":["pessoa-controller"],"operationId":"atualizarPessoa","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PessoaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PessoaDTO"}}}}}},"delete":{"tags":["pessoa-controller"],"operationId":"deletarPessoa","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/notificacoes/{id}":{"get":{"tags":["notificacao-controller"],"operationId":"buscarPorId_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificacaoDTO"}}}}}},"put":{"tags":["notificacao-controller"],"operationId":"atualizar_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificacaoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificacaoDTO"}}}}}},"delete":{"tags":["notificacao-controller"],"operationId":"deletar_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/tarefas":{"get":{"tags":["tarefa-controller"],"operationId":"listarTodas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TarefaDTO"}}}}}}},"post":{"tags":["tarefa-controller"],"operationId":"criar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TarefaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TarefaDTO"}}}}}}},"/salas":{"get":{"tags":["sala-controller"],"operationId":"listar","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SalaDTO"}}}}}}},"post":{"tags":["sala-controller"],"operationId":"criar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SalaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SalaDTO"}}}}}}},"/reunioes":{"get":{"tags":["reuniao-controller"],"operationId":"listar_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReuniaoDTO"}}}}}}},"post":{"tags":["reuniao-controller"],"operationId":"criar_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReuniaoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReuniaoDTO"}}}}}}},"/reunioes/{id}/presenca":{"post":{"tags":["presenca-controller"],"operationId":"registrarPresenca","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PresencaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PresencaDTO"}}}}}}},"/reunioes/{id}/encerrar":{"post":{"tags":["reuniao-controller"],"operationId":"encerrar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReuniaoDTO"}}}}}}},"/reunioes/testar-email":{"post":{"tags":["reuniao-controller"],"operationId":"testarEmail","parameters":[{"name":"email","in":"query","required":false,"schema":{"type":"string","default":"teste@exemplo.com"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/pessoas":{"get":{"tags":["pessoa-controller"],"operationId":"listarTodas_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PessoaDTO"}}}}}}},"post":{"tags":["pessoa-controller"],"operationId":"criarPessoa","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PessoaCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PessoaDTO"}}}}}}},"/notificacoes":{"get":{"tags":["notificacao-controller"],"operationId":"listar_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificacaoDTO"}}}}}}},"post":{"tags":["notificacao-controller"],"operationId":"criar_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificacaoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificacaoDTO"}}}}}}},"/auth/registro":{"post":{"tags":["auth-controller"],"operationId":"registro","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistroDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/tarefas/reuniao/{idReuniao}/pendencias":{"get":{"tags":["tarefa-controller"],"operationId":"verificarPendencias","parameters":[{"name":"idReuniao","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/relatorios/tarefas-concluidas":{"get":{"tags":["relatorio-controller"],"operationId":"getTarefasConcluidas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/relatorios/reunioes-por-sala":{"get":{"tags":["relatorio-controller"],"operationId":"getReunioesPorSala","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/relatorios/presenca-pessoa":{"get":{"tags":["relatorio-controller"],"operationId":"getPresencaPorPessoa","parameters":[{"name":"pessoaId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/relatorios/duracao-reunioes":{"get":{"tags":["relatorio-controller"],"operationId":"getDuracaoReunioes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/emails/teste":{"get":{"tags":["email-controller"],"operationId":"testarEmail_1","parameters":[{"name":"destinatario","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"TarefaDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"descricao":{"type":"string"},"prazo":{"type":"string","format":"date"},"concluida":{"type":"boolean"},"statusTarefa":{"type":"string","enum":["PRE_REUNIAO","POS_REUNIAO"]},"prioridade":{"type":"string"},"responsavelId":{"type":"integer","format":"int64"},"responsavelNome":{"type":"string"},"reuniaoId":{"type":"integer","format":"int64"}}},"SalaDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"capacidade":{"type":"integer","format":"int32"},"localizacao":{"type":"string"},"status":{"type":"string","enum":["LIVRE","OCUPADA","RESERVADA"]}}},"PessoaDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"email":{"type":"string"},"papel":{"type":"string","enum":["ADMIN","ORGANIZADOR","PARTICIPANTE"]},"crachaId":{"type":"string"}}},"ReuniaoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dataHoraInicio":{"type":"string","format":"date-time"},"duracaoMinutos":{"type":"integer","format":"int32"},"pauta":{"type":"string"},"ata":{"type":"string"},"salaId":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["AGENDADA","EM_ANDAMENTO","FINALIZADA","CANCELADA"]},"organizador":{"$ref":"#/components/schemas/PessoaDTO"},"sala":{"$ref":"#/components/schemas/SalaDTO"},"participantes":{"type":"array","items":{"$ref":"#/components/schemas/PessoaDTO"}},"participantesIds":{"type":"array","items":{"type":"integer","format":"int64"}},"dataHoraFim":{"type":"string","format":"date-time"}}},"NotificacaoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"mensagem":{"type":"string"},"dataEnvio":{"type":"string","format":"date-time"},"tipo":{"type":"string","enum":["EMAIL","CONSOLE","PUSH"]},"destinatarioId":{"type":"integer","format":"int64"}}},"PresencaDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"crachaId":{"type":"string"},"nomeParticipante":{"type":"string"},"horaEntrada":{"type":"string","format":"date-time"},"validadoPorCracha":{"type":"boolean"},"reuniaoId":{"type":"integer","format":"int64"}}},"PessoaCreateDTO":{"type":"object","properties":{"nome":{"type":"string"},"email":{"type":"string"},"papel":{"type":"string","enum":["ADMIN","ORGANIZADOR","PARTICIPANTE"]},"crachaId":{"type":"string"},"senha":{"type":"string"}}},"RegistroDTO":{"type":"object","properties":{"nome":{"type":"string"},"email":{"type":"string"},"senha":{"type":"string"},"papel":{"type":"string","enum":["ADMIN","ORGANIZADOR","PARTICIPANTE"]},"crachaId":{"type":"string"}}},"LoginDTO":{"type":"object","properties":{"email":{"type":"string"},"senha":{"type":"string"}}}},"securitySchemes":{"JWT":{"type":"http","description":"Insira o token JWT com o prefixo Bearer. Exemplo: 'Bearer abcde12345'","name":"JWT","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}